requirements.txt
attrs==20.3.0
fastapi==0.60.2
gunicorn==20.1.0
importlib-metadata==3.10.0
iniconfig==1.1.1
Mako==1.1.4
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pydantic==1.8.1
pyparsing==2.4.7
pytest==6.2.3
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
zipp==3.4.1
uvicorn==0.13.4
python-dotenv==0.17.0
aioredis==1.3.1
fastapi-plugins==0.3.0
requests==2.25.1
pytest-mock==3.5.1
prometheus-fastapi-instrumentator==5.7.1
websockets==8.1
arctic==1.79.4
tqdm==4.36.1
pandas==0.25.0

